CASMPET-4095: Stop using istio sidecar for these pods
The cray-precache-images pods don't do any communication over the
mesh so have no need for the istio-proxy sidecar. Not having the
sidecar makes it so that there's less to go wrong with this simple
service so it shouldn't have the sidecar.

This change configures the DaemonSet so that it tells the sidecar
injector to not include it.